solvelinmod: scale equation axes more if the problem is under-constrained
If the number of bits in the bounds exceeds the number of bits in the moduli, there
are likely to be many solutions. In that case, scale the equation axes a little more
to produce solutions anyway. Note that we generally want to avoid over-scaling the
equation axes for efficiency reasons.